+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Showing started at, last paused at and last resumed at for stopwatch/count down in the table 
+- CLI argument within "create" subcommand. Allows to specify a start date of the stopwatch or count down
 
 
 ### Added